+++ src/onegov.winterthur/onegov/winterthur/layout.py
@@ -3,6 +3,7 @@ from onegov.org.layout import DefaultLayout
 from onegov.core.elements import Link, Intercooler
 from onegov.winterthur import _
 from onegov.winterthur.collections import AddressCollection
+from onegov.winterthur.roadwork import RoadworkCollection
 
 
 class AddressLayout(DefaultLayout):
@@ -42,4 +43,25 @@ class AddressSubsetLayout(DefaultLayout):
             Link(_("Homepage"), self.homepage_url),
             Link(_("Addresses"), self.request.class_link(AddressCollection)),
             Link(_(self.model.street), '#')
+        ]
+
+
+class RoadworkCollectionLayout(DefaultLayout):
+
+    @cached_property
+    def breadcrumbs(self):
+        return [
+            Link(_("Homepage"), self.homepage_url),
+            Link(_("Roadworks"), '#'),
+        ]
+
+
+class RoadworkLayout(DefaultLayout):
+
+    @cached_property
+    def breadcrumbs(self):
+        return [
+            Link(_("Homepage"), self.homepage_url),
+            Link(_("Roadworks"), self.request.class_link(RoadworkCollection)),
+            Link(self.model.title, self.request.link(self.model))
         ]