+++ src/onegov.winterthur/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds a roadworks directory.
+  [href]
+
 0.3.5 (2019-02-28)
 ~~~~~~~~~~~~~~~~~~~~~